pyproject.toml
[tool.poetry]
name = "linux-collection-agent"
version = "0.1.0"
description = "Utility script to report hardware information and installed packages on a Linux machine."
authors = ["Bendik Samseth <bendik.samseth@inspera.no>"]

[tool.poetry.dependencies]
python = "^3.7"
google-api-python-client = "^1.12.8"
google-auth-httplib2 = "^0.0.4"
google-auth-oauthlib = "^0.4.2"
python-crontab = "^2.5.1"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
inspera-linux-collection-agent = 'linux_collection_agent.send_dump:send_dump'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
